Split up ondemand_dom_api_tests for sanitize build times
